Write code in parent object, and move common classes in common package (#19)
