--- src/main/java/APageParent.java
@@ -1,39 +1,0 @@
-
-import java.util.*;
-
-/**
- * Only for make diagram reading easier.
- */
-public abstract class APageParent {
-
-    protected HeaderFull headerFull = new HeaderFull();
-    /**
-     * Default constructor
-     */
-    public APageParent() {
-
-    }
-
-    //Create Header Simple
-    public Header createHeaderSimple(){
-        Header header = new Header();
-        return header;
-    }
-
-    //Create Header Full
-    public HeaderFull createHeader(){
-        HeaderFull header = new HeaderFull();
-        return header();
-    }
-
-
-
-
-
-
-
-
-
-
-
-}
