+++ src/main/java/Register/RegisterPage.java
@@ -1,5 +1,5 @@
 package Register;
-
+import java.*;
 import java.util.*;
 
 /**
